Verify itemset bindings for internal secondary instances
