+++ src/main/java/org/javarosa/xform/parse/FormInstanceParser.java
@@ -348,7 +348,7 @@ class FormInstanceParser {
 
             // Don't try to validate references if the external instance could not be resolved. We allow parsing a form
             // with placeholder external secondary instances for cases where a ReferenceManager can't be configured.
-            if (fi instanceof ExternalDataInstance && !((ExternalDataInstance) fi).isUsingPlaceholder()) {
+            if (!(fi instanceof ExternalDataInstance) || !((ExternalDataInstance) fi).isUsingPlaceholder()) {
                 if (fi.getTemplatePath(itemset.labelRef) == null) {
                     throw new XFormParseException("<label> node for itemset doesn't exist! [" + itemset.labelRef + "]");
                 }
