--- src/test/resources/org/javarosa/xform/parse/mixed-schema.xml
@@ -1,9 +1,0 @@
-<root>
-    <item>
-        <label>A</label>
-        <name>a</name>
-    </item>
-    <item>
-        <foo>bar</foo>
-    </item>
-</root>
