--- src/test/resources/org/javarosa/xform/parse/mixed-schema2.xml
@@ -1,9 +1,0 @@
-<root>
-    <item>
-        <foo>bar</foo>
-    </item>
-    <item>
-        <label>A</label>
-        <name>a</name>
-    </item>
-</root>
